refactor: update import paths for OrbitImages component
