+++ src/components/ui/orbit-images.tsx
@@ -5,7 +5,8 @@ import { useEffect, useRef, useState } from "react";
 
 import { cn } from "@/lib/utils";
 
-import { Button } from "./ui/button";
+import { Button } from "./button";
+
 
 interface OrbitImagesProps {
   title: string;
